Move the common function on in a specific module.
The is_valid_path and pluralize functions is not specific to the cli and
can be used by others things.
